Added in a new feature to sync checks using the check sync api
Exposed the check sync api to the dashboard/ui. This allows the user to
import the checks down from riak to the sensu master via the ui. Also
allows for the user to restart sensu (load the checks) via the ui. This
should be locked down by user later on.